Add files via upload
VFX: All of the screen tone overlays I'm messing with in ATL; figured we would have a few so I decided to organize the directory into folders. 
Emotes: exactly what you think; they'll go over top sprites here and there, but they're not essential enough to be included in the actual sprite psds. mostly putting them in a folder for the sake of organization.